Modifies the test suite to just test some jeopardy functions. Also modifies it so it will look for different test suites in the tests folder and run them separately.
