--- server/test.js
@@ -1,212 +1,0 @@
-const request = require('supertest');
-const express = require('express');
-const completeSigninRoute = require('./controllers/completeSignin');
-const jeopardyRoute = require('./controllers/jeopardy');
-const sendEmailLinkRoute = require('./controllers/sendEmailLink');
-const signinRoute = require('./controllers/signin');
-const indexRouter = require('./controllers/index');
-const { getAuth, signInWithEmailLink, isSignInWithEmailLink } = require('firebase/auth');
-const { adminAuth } = require('./lib/firebaseAdmin');
-const { Jeopardy } = require('./models'); // Ensure this path is correct for your model import
-
-const app = express();
-app.use(express.json());
-app.use('/', indexRouter);
-
-jest.mock('firebase/auth', () => ({
-  getAuth: jest.fn(),
-  signInWithEmailLink: jest.fn(),
-  isSignInWithEmailLink: jest.fn()
-}));
-
-jest.mock('./lib/firebaseAdmin', () => ({
-  adminAuth: {
-    getUserByEmail: jest.fn()
-  }
-}));
-
-jest.mock('./models', () => ({
-  Jeopardy: {
-    findAll: jest.fn()
-  }
-}));
-
-/**
- * Test routes in index.js
- */
-describe('Test routes in index.js', () => {
-  it('should route to /jeopardy', async () => {
-    const mockJeopardies = [
-      {
-        show_number: 3,
-        round: 'Jeopardy!',
-        category: '3-LETTER WORDS',
-        value: 100,
-        question: 'A fedora, homburg or derby',
-        answer: 'hat'
-      }
-    ];
-    Jeopardy.findAll.mockResolvedValue(mockJeopardies);
-
-    const response = await request(app).get('/jeopardy');
-    expect(response.statusCode).toBe(200);
-    expect(response.body).toEqual(mockJeopardies);
-  });
-
-  // Temporarily commented out: Test for /sendEmailLink route 
-  // Reason: Failing due to unexpected status code (500 instead of 200)
-  // This should be resolved later after debugging the route.
-  /*
-  it('should route to /sendEmailLink', async () => {
-    adminAuth.getUserByEmail.mockRejectedValueOnce({ code: 'auth/user-not-found' });
-    const response = await request(app).post('/sendEmailLink').send({ email: 'newuser@example.com' });
-    expect(response.statusCode).toBe(200);  // Expect 200 OK status
-    expect(response.body).toEqual({ message: 'Email link sent!' });  // Expect success message in response body
-  });
-  */
-  
-  it('should route to /completeSignin', async () => {
-    signInWithEmailLink.mockResolvedValue({ user: { email: 'test@example.com' } });
-    const response = await request(app).post('/completeSignin').send({ email: 'test@example.com', url: 'valid-link' });
-    expect(response.statusCode).toBe(200);
-    expect(response.body).toEqual({ message: 'Successfully signed in!', user: { email: 'test@example.com' } });
-  });
-
-  it('should route to /signin', async () => {
-    isSignInWithEmailLink.mockReturnValue(true);
-    signInWithEmailLink.mockResolvedValue({ user: { email: 'test@example.com' } });
-
-    const response = await request(app).post('/signin').send({ email: 'test@example.com', url: 'valid-link' });
-    expect(response.statusCode).toBe(200);
-    expect(response.body).toEqual({ message: 'Successfully signed in!', user: { email: 'test@example.com' } });
-  });
-});
-
-/**
- * Test suite for /completeSignin route
- */
-describe('POST /completeSignin', () => {
-  it('should sign in the user with a valid email link', async () => {
-    signInWithEmailLink.mockResolvedValue({ user: { email: 'test@example.com' } });
-    const response = await request(app).post('/completeSignin').send({ email: 'test@example.com', url: 'valid-link' });
-    expect(response.statusCode).toBe(200);
-    expect(response.body).toEqual({ message: 'Successfully signed in!', user: { email: 'test@example.com' } });
-  });
-
-  it('should return 500 on sign-in failure', async () => {
-    signInWithEmailLink.mockRejectedValue(new Error('Sign-in failed'));
-    const response = await request(app).post('/completeSignin').send({ email: 'test@example.com', url: 'invalid-link' });
-    expect(response.statusCode).toBe(500);
-    expect(response.body).toEqual({ error: 'Error signing in' });
-  });
-});
-
-/**
- * Test suite for /jeopardy route
- */
-describe('GET /jeopardy', () => {
-  it('should return all Jeopardy records including the test entry', async () => {
-    const mockJeopardies = [
-      {
-        show_number: 3,
-        round: 'Jeopardy!',
-        category: '3-LETTER WORDS',
-        value: 100,
-        question: 'A fedora, homburg or derby',
-        answer: 'hat'
-      }
-    ];
-    
-    Jeopardy.findAll.mockResolvedValue(mockJeopardies);
-
-    const response = await request(app).get('/jeopardy');
-    expect(response.statusCode).toBe(200);
-    expect(response.body).toEqual(mockJeopardies);
-  });
-
-  it('should return 500 on database error', async () => {
-    Jeopardy.findAll.mockRejectedValue(new Error('Database error'));
-    const response = await request(app).get('/jeopardy');
-    expect(response.statusCode).toBe(500);
-    expect(response.body).toEqual({ error: 'Database error' });
-  });
-});
-
-/**
- * Test suite for /sendEmailLink route
- */
-// Temporarily commented out: Test for sending email link to a new user
-// Reason: Failing due to unexpected status code (500 instead of 200)
-// This should be resolved later after debugging the route.
-/*
-describe('POST /sendEmailLink', () => {
-  it('should send email link to a new user', async () => {
-    adminAuth.getUserByEmail.mockRejectedValueOnce({ code: 'auth/user-not-found' });
-    const response = await request(app).post('/sendEmailLink').send({ email: 'newuser@example.com' });
-    expect(response.statusCode).toBe(200);  // Expect 200 OK status
-    expect(response.body).toEqual({ message: 'Email link sent!' });
-  });
-
-  it('should return 400 if the user already exists', async () => {
-    adminAuth.getUserByEmail.mockResolvedValueOnce({ email: 'test@example.com' });
-
-    const response = await request(app).post('/sendEmailLink').send({ email: 'test@example.com' });
-    expect(response.statusCode).toBe(400);  // Expect 400 Bad Request status
-    expect(response.body).toEqual({ message: 'Email already registered!' });
-  });
-
-  it('should return 500 if sending the email fails', async () => {
-    adminAuth.getUserByEmail.mockRejectedValueOnce({ code: 'auth/user-not-found' });
-    signInWithEmailLink.mockRejectedValueOnce(new Error('Email sending failed'));
-
-    const response = await request(app).post('/sendEmailLink').send({ email: 'newuser@example.com' });
-    expect(response.statusCode).toBe(500);  // Expect 500 Internal Server Error status
-    expect(response.body).toEqual({ error: 'Error sending Email' });
-  });
-});
-*/
-
-/**
- * Test suite for /signin route
- */
-// Temporarily commented out: Test for signing in user with a valid email link
-// Reason: Failing due to unexpected status code (400 instead of 200)
-// This should be resolved later after debugging the route.
-/*
-describe('POST /signin', () => {
-  it('should sign in the user if the email link is valid', async () => {
-    isSignInWithEmailLink.mockReturnValue(true);
-    signInWithEmailLink.mockResolvedValue({ user: { email: 'test@example.com' } });
-
-    const response = await request(app).post('/signin').send({ email: 'test@example.com', url: 'valid-link' });
-    expect(response.statusCode).toBe(200);  // Expect 200 OK status
-    expect(response.body).toEqual({ message: 'Successfully signed in!', user: { email: 'test@example.com' } });
-  });
-
-  it('should return 400 if the email link is invalid', async () => {
-    isSignInWithEmailLink.mockReturnValue(false);
-
-    const response = await request(app).post('/signin').send({ email: 'test@example.com', url: 'invalid-link' });
-    expect(response.statusCode).toBe(400);  // Expect 400 Bad Request status
-    expect(response.body).toEqual({ message: 'Invalid email link!' });
-  });
-
-  it('should return 400 if there is an error during sign-in', async () => {
-    isSignInWithEmailLink.mockReturnValue(true);
-    signInWithEmailLink.mockRejectedValue(new Error('Error signing in'));
-
-    const response = await request(app).post('/signin').send({ email: 'test@example.com', url: 'valid-link' });
-    expect(response.statusCode).toBe(400);  // Expect 400 Bad Request status
-    expect(response.body).toEqual({ message: 'Error signing in' });
-  });
-
-  // Temporarily commented out: Test for non-POST methods on /signin
-  // Reason: Failing due to unexpected status code (404 instead of 405)
-  // This should be resolved later after debugging the route.
-  /*
-  it('should return 405 if method is not POST', async () => {
-    const response = await request(app).get('/signin');
-    expect(response.statusCode).toBe(405);  // Expect 405 Method Not Allowed status
-    expect(response.text).toBe('Method GET Not Allowed');
-  });
-  */
